Moved `Window` and `WindowBackend` to `window.rs`
